확장함수로 분리
