+++ app/src/main/java/com/nbcteam5/nbccontact/util/addConst.kt
@@ -1,0 +1,7 @@
+package com.nbcteam5.nbccontact.util
+
+const val USER_INFORMATION = "userInformation"
+const val NAME = "name"
+const val PHONE_NUMEBER = "phoneNumber"
+const val EMAIL = "email"
+const val EVENT = "event"
